pyproject.toml
[project]
name = "MathMate"
version = "0.1.0"
description = "Multi-agent workflow for answering math questions with LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "pydantic>=2.7",
    "Pillow>=10.0",
    "pytesseract>=0.3.10",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.20",
    "faiss-cpu>=1.12.0",
    "numpy>=2.2.6",
    "dashscope>=1.24.6",
    "pytest>=8.4.2",
    "openai>=2.3.0",
    "copilotkit>=0.1.69",
    "ag-ui-langgraph>=0.0.17",
    "python-dotenv>=1.0.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

filterwarnings = [
  "ignore:.*SwigPy.*:DeprecationWarning",
  "ignore:.*no __module__ attribute:DeprecationWarning",
]
